Move script tests out of 'interpreter' tests
